python/requirements.txt
supabase==2.4.2
httpx==0.27.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
resend==0.8.0
schedule==1.2.1
prophet==1.1.5
redis==5.0.0
